add: predict integration and new home page layout

+++ lib/app/data/models/predict_model.dart
@@ -122,7 +122,7 @@ class DetectedObject {
 
   factory DetectedObject.fromJson(Map<String, dynamic> json) => DetectedObject(
         name: json["name"],
-        detectedObjectClass: json["class"],
+        detectedObjectClass: json["class"].toInt(),
         point: json["point"],
       );
